pyproject.toml
[tool.poetry]
name = "job_scrp"
version = "0.1.0"
description = "project for sraping data about spk swimming pool and sauna capacity"
authors = ["Jakub Turner <kuba.turner@gmail.com>"]
packages = [
    {include="spk_scrp", from="src"}
]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
mysql-connector-python = "^8.0.31"
psycopg2-binary = "2.9.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
spk_scrp = "spk_scrp.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "avast-ida-remote"
url = "https://artifactory.ida.avast.com/artifactory/api/pypi/pypi-remote/simple/"
default = true

[[tool.poetry.source]]
name = "avast-ida-local"
url = "https://artifactory.ida.avast.com/artifactory/api/pypi/pypi-local/simple/"